// src/app/expense-category.enum.ts
export enum ExpenseCategory {
  Food = 'Food',
  Transport = 'Transport',
  Bills = 'Bills',
  Entertainment = 'Entertainment',
  Shopping = 'Shopping',
  Healthcare = 'Healthcare',
  Education = 'Education',
  Travel = 'Travel',
  Gifts = 'Gifts',
  Utilities = 'Utilities',
  Insurance = 'Insurance',
  Other = 'Other'
}